refactor all moving parts to stratosphere directory
