+++ esmvaltool/diag_scripts/stratosphere/area_base.py
