+++ esmvaltool/recipes/recipe_autoassess_landsurface_permafrost.yml
@@ -8,7 +8,7 @@ documentation:
     Climatological files are stored externally to avoid overloading the
     ESMValTool source. See /gws/nopw/j04/esmeval/autoassess_specific_files
     (on JASMIN).
-    
+
   authors:
     - predoi_valeriu
     - sellar_alistair
@@ -24,8 +24,9 @@ documentation:
     - predoi_valeriu
 
 datasets:
-    - {dataset: ACCESS-CM2,  project: CMIP6, exp: historical,  grid: gn,  ensemble: r1i1p1f1,  start_year: 1992,  end_year: 2002}
-    - {dataset: E3SM-1-0,  project: CMIP6, exp: historical,  grid: gr,  ensemble: r1i1p1f1,  start_year: 1992,  end_year: 2002}
+  - {dataset: ACCESS-CM2, project: CMIP6, exp: historical, grid: gn, ensemble: r1i1p1f1, start_year: 1992,
+     end_year: 2002}
+  - {dataset: E3SM-1-0, project: CMIP6, exp: historical, grid: gr, ensemble: r1i1p1f1, start_year: 1992, end_year: 2002}
 
 preprocessors:
   pp_aa_area:
@@ -47,7 +48,7 @@ diagnostics:
         mip: Lmon
     scripts:
       autoassess_landsurf_permafrost: &autoassess_landsurf_permafrost_settings
-        script: autoassess/autoassess_area_base.py
+        script: autoassess/area_base.py
         title: "Autoassess Land-Surface Permafrost Diagnostic"
         area: land_surface_permafrost
         control_model: E3SM-1-0
@@ -64,7 +65,7 @@ diagnostics:
         <<: *autoassess_landsurf_permafrost_settings
         control_model: E3SM-1-0
         exp_model: ACCESS-CM2
-        script: autoassess/plot_autoassess_metrics.py
+        script: autoassess/plot_metrics.py
         ancestors: ['*/autoassess_landsurf_permafrost']
         title: "Plot Land-Surface Permafrost Metrics"
         plot_name: "Permafrost_Metrics"
