+++ esmvaltool/recipes/recipe_autoassess_landsurface_soilmoisture.yml
@@ -23,8 +23,9 @@ documentation:
     - predoi_valeriu
 
 datasets:
-    - {dataset: ACCESS-CM2,  project: CMIP6, exp: historical,  grid: gn,  ensemble: r1i1p1f1,  start_year: 1992,  end_year: 2002}
-    - {dataset: E3SM-1-0,  project: CMIP6, exp: historical,  grid: gr,  ensemble: r1i1p1f1,  start_year: 1992,  end_year: 2002}
+  - {dataset: ACCESS-CM2, project: CMIP6, exp: historical, grid: gn, ensemble: r1i1p1f1, start_year: 1992,
+     end_year: 2002}
+  - {dataset: E3SM-1-0, project: CMIP6, exp: historical, grid: gr, ensemble: r1i1p1f1, start_year: 1992, end_year: 2002}
 
 preprocessors:
   pp_aa_area:
@@ -36,11 +37,11 @@ diagnostics:
   aa_landsurf_soilmoisture:
     description: Autoassess test diag for Land-Surface Soilmoisture.
     variables:
-      mrsos: # moisture_content_of_soil_layer
+      mrsos:  # moisture_content_of_soil_layer
         mip: Lmon
     scripts:
       autoassess_landsurf_soilmoisture: &autoassess_landsurf_soilmoisture_settings
-        script: autoassess/autoassess_area_base.py
+        script: autoassess/area_base.py
         title: "Autoassess Land-Surface Soilmoisture Diagnostic"
         area: land_surface_soilmoisture
         control_model: ACCESS-CM2
@@ -57,7 +58,7 @@ diagnostics:
         <<: *autoassess_landsurf_soilmoisture_settings
         control_model: ACCESS-CM2
         exp_model: E3SM-1-0
-        script: autoassess/plot_autoassess_metrics.py
+        script: autoassess/plot_metrics.py
         ancestors: ['*/autoassess_landsurf_soilmoisture']
         title: "Plot Land-Surface Soilmoisture Metrics"
         plot_name: "Soilmoisture_Metrics"
